fix: resolve build errors and implement missing kernel/userspace functionality
 - fix(d3d): resolve syntax errors and implement workload tracking in D3D wrapper\n - fix(sycl): use correct unmap function in memory free\n - fix(kernel): implement PCI probe/remove and fix P2P detection\n - fix(build): integrate tools into CMake and fix generator expression issues\n - docs: add comprehensive documentation for architecture, API, and installation

+++ kernel/vendors/mvgal_nvidia.h
@@ -84,7 +84,7 @@ int mvgal_nvidia_query_utilization(struct mvgal_gpu_device *gpu, uint32_t *util_
 /* Device detection helpers */
 u64 mvgal_nvidia_estimate_vram(u16 device_id);
 u32 mvgal_nvidia_estimate_bandwidth(u16 device_id);
-float mvgal_nvidia_get_compute_capability(u16 device_id);
+u32 mvgal_nvidia_get_compute_capability(u16 device_id);
 u32 mvgal_nvidia_get_sm_count(u16 device_id);
 bool mvgal_nvidia_is_rtx_series(u16 device_id);
 bool mvgal_nvidia_has_nvlink(struct pci_dev *pdev);
